Integracion vuelo paquete turistico, hotel + api responses y exceptions

+++ src/main/java/com/viajaYa/viajaYa/repositories/IPaqueteTuristicoRepository.java
@@ -2,9 +2,9 @@ package com.viajaYa.viajaYa.repositories;
 
 import com.viajaYa.viajaYa.models.PaqueteTuristicoModel;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPaqueteTuristicoRepository extends JpaRepository<PaqueteTuristicoModel, Integer> {
-
+public interface IPaqueteTuristicoRepository extends JpaRepository<PaqueteTuristicoModel, Long> {
 }
